Move file contents & rename stuff.
