+++ include/aws/http/request_response.h
@@ -1,5 +1,5 @@
-#ifndef AWS_HTTP_MESSAGE_H
-#define AWS_HTTP_MESSAGE_H
+#ifndef AWS_HTTP_REQUEST_RESPONSE_H
+#define AWS_HTTP_REQUEST_RESPONSE_H
 
 /*
  * Copyright 2010-2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
@@ -19,24 +19,29 @@
 #include <aws/http/http.h>
 
 struct aws_http_connection;
-struct aws_http_exchange;
+
+/**
+ * A stream exists for the duration of a request/response exchange.
+ * A client creates a stream to send a request and receive a response.
+ * A server creates a stream to receive a request and send a response.
+ * In http/2, a push-promise stream can be sent by a server and received by a client.
+ */
+struct aws_http_stream;
 
 /**
  * Called repeatedly whenever body data can be sent.
  * Write up to `size` bytes to the `buffer` and return the number of bytes written.
  * Return 0 to end the body.
  */
-typedef size_t(
-    aws_http_body_sender_fn)(struct aws_http_exchange *exchange, uint8_t *buffer, size_t size, void *user_data);
+typedef size_t(aws_http_body_sender_fn)(struct aws_http_stream *stream, uint8_t *buffer, size_t size, void *user_data);
 
 typedef void(aws_http_on_incoming_headers_fn)(
-    struct aws_http_exchange *exchange,
+    struct aws_http_stream *stream,
     const struct aws_http_header *header_array,
     size_t num_headers,
     void *user_data);
 
-typedef void(
-    aws_http_on_incoming_header_block_done_fn)(struct aws_http_exchange *exchange, bool has_body, void *user_data);
+typedef void(aws_http_on_incoming_header_block_done_fn)(struct aws_http_stream *stream, bool has_body, void *user_data);
 
 /**
  * Called repeatedly as body data is received.
@@ -45,15 +50,15 @@ typedef void(
  * `out_window_update_size` is the amount by which the window is updated after reading data.
  * By default, it is the same size as the data which has just come in.
  * To prevent the window from updating, set this value to 0.
- * The window can be manually updated later via aws_http_exchange_update_window()
+ * The window can be manually updated later via aws_http_stream_update_window()
  */
 typedef void(aws_http_on_incoming_body_fn)(
-    struct aws_http_exchange *exchange,
+    struct aws_http_stream *stream,
     const struct aws_byte_cursor *data,
     size_t *out_window_update_size,
     void *user_data);
 
-typedef void(aws_http_on_exchange_complete_fn)(struct aws_http_exchange *exchange, int error_code, void *user_data);
+typedef void(aws_http_on_stream_complete_fn)(struct aws_http_stream *stream, int error_code, void *user_data);
 
 struct aws_http_request_options {
     /**
@@ -103,7 +108,7 @@ struct aws_http_request_options {
 
     /**
      * Invoked repeatedly times as headers are received.
-     * At this point, aws_http_exchange_get_incoming_response_status() can be called.
+     * At this point, aws_http_stream_get_incoming_response_status() can be called.
      * Optional.
      */
     aws_http_on_incoming_headers_fn *on_response_headers;
@@ -121,10 +126,10 @@ struct aws_http_request_options {
     aws_http_on_incoming_body_fn *on_response_body;
 
     /**
-     * Invoked when request/response exchange is complete, whether successful or unsucessful
+     * Invoked when request/response stream is complete, whether successful or unsucessful
      * Optional.
      */
-    aws_http_on_exchange_complete_fn *on_complete;
+    aws_http_on_stream_complete_fn *on_complete;
 };
 
 struct aws_http_request_handler_options {
@@ -137,7 +142,7 @@ struct aws_http_request_handler_options {
     aws_http_on_incoming_headers_fn *on_request_headers;
     aws_http_on_incoming_header_block_done_fn *on_request_header_block_done;
     aws_http_on_incoming_body_fn *on_request_body;
-    aws_http_on_exchange_complete_fn *on_complete;
+    aws_http_on_stream_complete_fn *on_complete;
 };
 
 struct aws_http_response_options {
@@ -153,57 +158,52 @@ struct aws_http_response_options {
 AWS_EXTERN_C_BEGIN
 
 /**
- * Create an exchange, with a client connection sending a request.
- * The request starts sending automatically once the exchange is created.
+ * Create a stream, with a client connection sending a request.
+ * The request starts sending automatically once the stream is created.
  * The `def`, and all memory it references, is copied during this call.
  */
 AWS_HTTP_API
-struct aws_http_exchange *aws_http_exchange_new_request(const struct aws_http_request_options *options);
+struct aws_http_stream *aws_http_stream_new_client_request(const struct aws_http_request_options *options);
 
 /**
- * Create an exchange, with a server connection receiving and responding to a request.
- * aws_http_exchange_send_response() should be used to send a response.
+ * Create a stream, with a server connection receiving and responding to a request.
+ * aws_http_stream_send_response() should be used to send a response.
  */
 AWS_HTTP_API
-struct aws_http_exchange *aws_http_exchange_new_request_handler(const struct aws_http_request_handler_options *options);
+struct aws_http_stream *aws_http_stream_new_server_request_handler(
+    const struct aws_http_request_handler_options *options);
 
 AWS_HTTP_API
-void aws_http_exchange_destroy(struct aws_http_exchange *exchange);
+void aws_http_stream_destroy(struct aws_http_stream *stream);
 
 AWS_HTTP_API
-struct aws_http_connection *aws_http_exchange_get_connection(const struct aws_http_exchange *exchange);
+struct aws_http_connection *aws_http_stream_get_connection(const struct aws_http_stream *stream);
 
-/* Only valid in "request" exchanges, once response headers start arriving */
+/* Only valid in "request" streams, once response headers start arriving */
 AWS_HTTP_API
-int aws_http_exchange_get_incoming_response_status(const struct aws_http_exchange *exchange, int *out_status);
+int aws_http_stream_get_incoming_response_status(const struct aws_http_stream *stream, int *out_status);
 
-/* Only valid in "request handler" exchanges, once request headers start arriving */
+/* Only valid in "request handler" streams, once request headers start arriving */
 AWS_HTTP_API
-int aws_http_exchange_get_incoming_request_method(
-    const struct aws_http_exchange *exchange,
-    enum aws_http_method *out_method);
+int aws_http_stream_get_incoming_request_method(const struct aws_http_stream *stream, enum aws_http_method *out_method);
 
 AWS_HTTP_API
-int aws_http_exchange_get_incoming_request_method_str(
-    const struct aws_http_exchange *exchange,
+int aws_http_stream_get_incoming_request_method_str(
+    const struct aws_http_stream *stream,
     struct aws_byte_cursor *out_method);
 
 AWS_HTTP_API
-int aws_http_exchange_get_incoming_request_uri(
-    const struct aws_http_exchange *exchange,
-    struct aws_byte_cursor *out_uri);
+int aws_http_stream_get_incoming_request_uri(const struct aws_http_stream *stream, struct aws_byte_cursor *out_uri);
 
-/* only callable from "request handler" exchanges */
+/* only callable from "request handler" streams */
 AWS_HTTP_API
-int aws_http_exchange_send_response(
-    struct aws_http_exchange *exchange,
-    const struct aws_http_response_options *options);
+int aws_http_stream_send_response(struct aws_http_stream *stream, const struct aws_http_response_options *options);
 
 /* Manually issue a window update.
  * This should only be called if the body reader is reducing the automatic window update size */
 AWS_HTTP_API
-int aws_http_exchange_update_window(struct aws_http_exchange *exchange, size_t increment_size);
+int aws_http_stream_update_window(struct aws_http_stream *stream, size_t increment_size);
 
 AWS_EXTERN_C_END
 
-#endif /* AWS_HTTP_MESSAGE_H */
+#endif /* AWS_HTTP_REQUEST_RESPONSE_H */
